% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesh.R
\name{check_sesh}
\alias{check_sesh}
\title{Check current conditions against a sesh.}
\usage{
check_sesh(path)
}
\arguments{
\item{path}{Valid path to a sesh saved CSV.}
}
\description{
The workhorse that checks loaded and installed pacakge versions againstst
those specified in sesh.
}
