% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesh.R
\name{convert_session_info}
\alias{convert_session_info}
\title{Convert a saved RDS with results of `devtools::session_info()` to a
sesh CSV.}
\usage{
convert_session_info(path)
}
\arguments{
\item{path}{Valid path to a `devtools::session_info()` output
saved as RDS.}
}
\description{
Opens up access to using other functions for loading critical packages.
}
