% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesh.R
\name{.check_installed}
\alias{.check_installed}
\title{a hidden function to search installed versions}
\usage{
.check_installed(package, version)
}
\arguments{
\item{package}{name of package}

\item{version}{version of the package}
}
\description{
a hidden function to search installed versions
}
