% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesh.R
\name{.extract_result}
\alias{.extract_result}
\title{a hidden function to parse return from 'purrr::safely()'}
\usage{
.extract_result(x)
}
\arguments{
\item{x}{the return of an attempted install wrapped in `safely()`}
}
\description{
a hidden function to parse return from 'purrr::safely()'
}
