% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesh.R
\name{.extract_sesh}
\alias{.extract_sesh}
\title{Extract a data frame of attached packages}
\usage{
.extract_sesh(session_info)
}
\arguments{
\item{session_info}{Output from `devtools::session_info()`}
}
\description{
Extract a data frame of attached packages
}
