% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesh.R
\name{.print_capture}
\alias{.print_capture}
\title{a hidden function to nicely print data from the console
from: https://stackoverflow.com/questions/26083625/how-do-you-include-data-frame-output-inside-warnings-and-errors}
\usage{
.print_capture(x)
}
\arguments{
\item{x}{expression to print}
}
\description{
a hidden function to nicely print data from the console
from: https://stackoverflow.com/questions/26083625/how-do-you-include-data-frame-output-inside-warnings-and-errors
}
