% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesh.R
\name{load_sesh}
\alias{load_sesh}
\title{Function to load out a sesh if required versions are installed.}
\usage{
load_sesh(path, interactive = TRUE)
}
\arguments{
\item{path}{Valid path to a sesh saved CSV.}

\item{interactive}{Returns a data frame of required installs if FALSE.}
}
\description{
Function to load out a sesh if required versions are installed.
}
