% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesh.R
\name{read_sesh}
\alias{read_sesh}
\title{Read a saved CSV to see critical package info.}
\usage{
read_sesh(path)
}
\arguments{
\item{path}{Valid path to a sesh saved CSV.}
}
\description{
Read a saved CSV to see critical package info.
}
