% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesh.R
\name{save_sesh}
\alias{save_sesh}
\title{Save `sesh` as csv.}
\usage{
save_sesh(path = "sesh_{as.character(Sys.Date())}.csv")
}
\arguments{
\item{path}{Valid path to a write a sesh saved CSV.}
}
\value{
Writes a CSV file with essential package information.
}
\description{
Save `sesh` as csv.
}
\examples{
save_sesh()
}
