% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesh.R
\name{sesh}
\alias{sesh}
\title{See info about attached R packages during an R session}
\usage{
sesh()
}
\value{
A data frame.
}
\description{
A lighter version \code{devtools::session_info()} that returns a dataframe
and only \emph{attached} libraries
}
