% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesh.R
\name{unload_sesh}
\alias{unload_sesh}
\title{Unload all sesh packages}
\usage{
unload_sesh(path)
}
\arguments{
\item{path}{A character. Valid file path to a `save_sesh()` output CSV.}
}
\description{
Useful for going back to gloabal versions when leaving a sesh.
}
