% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesh.R
\name{install_sesh}
\alias{install_sesh}
\title{A function to install specific package versions.}
\usage{
install_sesh(path, auth_token = devtools::github_pat(quiet), ...)
}
\arguments{
\item{path}{A character. Valid file path to a \code{save_sesh()} output CSV.}

\item{auth_token}{A character. Passed to \code{devtools::install_github()} for accessing
private repos.}

\item{...}{Arguments passed to \code{devtools::install()}.}
}
\description{
Works for packages installed from CRAN and GitHub.
}
