% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesh.R
\name{save_sesh}
\alias{save_sesh}
\title{Save `sesh` output as csv.}
\usage{
save_sesh(path = "sesh_{as.character(Sys.Date())}.RDS")
}
\value{
Saves a RDS file with essential package information.
}
\description{
Renames columns from `sesh()` to work `sesh_check()`.
}
\examples{
save_sesh()
}
