% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesh.R
\name{save_sesh}
\alias{save_sesh}
\title{Save `sesh` output as csv.}
\usage{
save_sesh(path = "sesh_{as.character(Sys.Date())}.csv")
}
\value{
Saves a CSV with essential information about loaded packages.
}
\description{
Works via `devtools::session_info()`.
}
\examples{
save_sesh()
}
